test: mobileの未実行テストをtests配下へ移動 #653
