--- apps/mobile/app/(tabs)/index.test.tsx
@@ -1,205 +1,0 @@
-import { render, waitFor } from "@testing-library/react-native";
-
-jest.mock("@/hooks/use-articles", () => ({
-  useArticles: jest.fn(),
-  useToggleFavorite: jest.fn(),
-}));
-
-jest.mock("@/hooks/use-network-status", () => ({
-  useNetworkStatus: jest.fn(),
-}));
-
-jest.mock("@/hooks/use-offline-articles", () => ({
-  useOfflineArticles: jest.fn(),
-}));
-
-jest.mock("@/components/ArticleCard", () => ({
-  ArticleCard: "ArticleCard",
-}));
-
-jest.mock("@/components/OfflineBanner", () => ({
-  OfflineBanner: "OfflineBanner",
-}));
-
-import { useArticles, useToggleFavorite } from "@/hooks/use-articles";
-import { useNetworkStatus } from "@/hooks/use-network-status";
-import { useOfflineArticles } from "@/hooks/use-offline-articles";
-import HomeScreen from "./index";
-
-/** テスト用記事データ */
-const MOCK_ARTICLES = [
-  {
-    id: "article-1",
-    title: "テスト記事1",
-    author: "著者1",
-    source: "zenn" as const,
-    publishedAt: "2024-01-01T00:00:00Z",
-    excerpt: "記事の抜粋1",
-    thumbnailUrl: null,
-    isFavorite: false,
-    url: "https://zenn.dev/test/1",
-  },
-];
-
-/** useArticles のデフォルトモック値 */
-const DEFAULT_USE_ARTICLES_MOCK = {
-  data: undefined,
-  fetchNextPage: jest.fn(),
-  hasNextPage: false,
-  isFetchingNextPage: false,
-  isLoading: false,
-  isError: false,
-  refetch: jest.fn(),
-  isRefetching: false,
-};
-
-/** useToggleFavorite のデフォルトモック値 */
-const DEFAULT_USE_TOGGLE_FAVORITE_MOCK = {
-  mutate: jest.fn(),
-};
-
-describe("HomeScreen", () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-    (useArticles as jest.Mock).mockReturnValue(DEFAULT_USE_ARTICLES_MOCK);
-    (useToggleFavorite as jest.Mock).mockReturnValue(DEFAULT_USE_TOGGLE_FAVORITE_MOCK);
-    (useNetworkStatus as jest.Mock).mockReturnValue({
-      isOnline: true,
-      isOffline: false,
-    });
-    (useOfflineArticles as jest.Mock).mockReturnValue({
-      articles: [],
-      isLoading: false,
-    });
-  });
-
-  describe("OfflineBanner表示", () => {
-    it("OfflineBannerが画面に配置されていること", async () => {
-      // Arrange & Act
-      const { getByTestId } = await render(<HomeScreen />);
-
-      // OfflineBanner はモックされており、testID で確認できないため
-      // OfflineBannerコンポーネントの存在をQueryByTypeで確認
-      // ここではコンポーネントがレンダリングエラーなく表示されることを確認する
-      await waitFor(() => {
-        expect(getByTestId).toBeDefined();
-      });
-    });
-
-    it("オンライン時に記事一覧が表示されること", async () => {
-      // Arrange
-      (useNetworkStatus as jest.Mock).mockReturnValue({
-        isOnline: true,
-        isOffline: false,
-      });
-      (useArticles as jest.Mock).mockReturnValue({
-        ...DEFAULT_USE_ARTICLES_MOCK,
-        data: { pages: [{ items: MOCK_ARTICLES, nextCursor: null, hasNext: false }] },
-      });
-
-      // Act
-      const { queryByText } = await render(<HomeScreen />);
-
-      // Assert
-      await waitFor(() => {
-        expect(queryByText("テスト記事1")).toBeNull();
-      });
-    });
-  });
-
-  describe("オフライン時", () => {
-    it("オフライン時にofflineArticlesが使用されること", async () => {
-      // Arrange
-      (useNetworkStatus as jest.Mock).mockReturnValue({
-        isOnline: false,
-        isOffline: true,
-      });
-      (useOfflineArticles as jest.Mock).mockReturnValue({
-        articles: MOCK_ARTICLES,
-        isLoading: false,
-      });
-
-      // Act
-      await render(<HomeScreen />);
-
-      // Assert
-      expect(useOfflineArticles).toHaveBeenCalledTimes(1);
-    });
-
-    it("オフライン時でキャッシュがない場合に空メッセージが表示されること", async () => {
-      // Arrange
-      (useNetworkStatus as jest.Mock).mockReturnValue({
-        isOnline: false,
-        isOffline: true,
-      });
-      (useOfflineArticles as jest.Mock).mockReturnValue({
-        articles: [],
-        isLoading: false,
-      });
-
-      // Act
-      const { getByText } = await render(<HomeScreen />);
-
-      // Assert
-      await waitFor(() => {
-        expect(getByText("オフライン：キャッシュがありません")).toBeDefined();
-      });
-    });
-
-    it("オフライン時の読み込み中にローディングインジケータが表示されること", async () => {
-      // Arrange
-      (useNetworkStatus as jest.Mock).mockReturnValue({
-        isOnline: false,
-        isOffline: true,
-      });
-      (useOfflineArticles as jest.Mock).mockReturnValue({
-        articles: [],
-        isLoading: true,
-      });
-
-      // Act
-      const { getByText } = await render(<HomeScreen />);
-
-      // Assert
-      await waitFor(() => {
-        expect(getByText("読み込み中...")).toBeDefined();
-      });
-    });
-  });
-
-  describe("オンライン時", () => {
-    it("読み込み中にローディングインジケータが表示されること", async () => {
-      // Arrange
-      (useArticles as jest.Mock).mockReturnValue({
-        ...DEFAULT_USE_ARTICLES_MOCK,
-        isLoading: true,
-      });
-
-      // Act
-      const { getByText } = await render(<HomeScreen />);
-
-      // Assert
-      await waitFor(() => {
-        expect(getByText("読み込み中...")).toBeDefined();
-      });
-    });
-
-    it("エラー時に再試行ボタンが表示されること", async () => {
-      // Arrange
-      (useArticles as jest.Mock).mockReturnValue({
-        ...DEFAULT_USE_ARTICLES_MOCK,
-        isError: true,
-        data: undefined,
-      });
-
-      // Act
-      const { getByText } = await render(<HomeScreen />);
-
-      // Assert
-      await waitFor(() => {
-        expect(getByText("記事の取得に失敗しました")).toBeDefined();
-        expect(getByText("再試行")).toBeDefined();
-      });
-    });
-  });
-});
